// supabase/types.d.ts
export interface NoteRow {
  id: number;
  title: string;
  content: string;
  tags: string[];
  created_at: string;
  updated_at: string;
  user_id: string;
  guild_id: string;
}
export interface NoteInsert {
  title: string;
  content: string;
  tags?: string[];
  created_at?: string;
  updated_at?: string;
  user_id: string;
  guild_id: string;
}
export interface NoteUpdate {
  title?: string;
  content?: string;
  tags?: string[];
  updated_at?: string;
}